+++ generic/MSECriterion.cu
@@ -13,7 +13,7 @@ void THNN_(MSECriterion_updateOutput)(
   THCUNN_check_dim_size(state, output, 1, 0, 1);
   THCUNN_assertSameGPU(state, 2, input, target);
 
-  long size = THCTensor_(nElement)(state, input);
+  ptrdiff_t size = THCTensor_(nElement)(state, input);
 
   input = THCTensor_(newContiguous)(state, input);
   target = THCTensor_(newContiguous)(state, target);
@@ -46,7 +46,7 @@ void THNN_(MSECriterion_updateGradInput)(
   THCUNN_check_nElement(state, input, target);
   THCUNN_assertSameGPU(state, 3, input, target, gradInput);
 
-  long size = THCTensor_(nElement)(state, input);
+  ptrdiff_t size = THCTensor_(nElement)(state, input);
   accreal norm = sizeAverage ? (accreal)(2)/size : (accreal)(2);
 
   input = THCTensor_(newContiguous)(state, input);
