+++ generic/SoftMarginCriterion.cu
@@ -14,7 +14,7 @@ void THNN_(SoftMarginCriterion_updateOutput)(
   THCUNN_assertSameGPU(state, 2, input, target);
   accreal sum;
 
-  long size = THCTensor_(nElement)(state, input);
+  ptrdiff_t size = THCTensor_(nElement)(state, input);
 
   input = THCTensor_(newContiguous)(state, input);
   target = THCTensor_(newContiguous)(state, target);
@@ -42,7 +42,7 @@ void THNN_(SoftMarginCriterion_updateGradInput)(
   THCUNN_check_nElement(state, input, target);
   THCUNN_assertSameGPU(state, 3, input, target, gradInput);
 
-  long size = THCTensor_(nElement)(state, input);
+  ptrdiff_t size = THCTensor_(nElement)(state, input);
   accreal norm = (sizeAverage ? 1./size : 1.);
 
   input = THCTensor_(newContiguous)(state, input);
